requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-JWT-Extended==4.4.4
Flask-Cors==3.0.10
Werkzeug==2.2.2
gunicorn==20.1.0